change game function
