+++ client/src/components/game/GameDashboard.jsx
@@ -10,11 +10,13 @@ import { UPDATE_GAMEDATA } from '../../utils/mutations';
 import { QUERY_USER } from '../../utils/queries';
 import { useParams } from 'react-router-dom';
 import auth from '../../utils/auth';
+import useGameHook from '../../lib/useGameHook';
 
 export default function GameDashboard() {
-  const { petState, setGameState, setPetState } = useGameContext();
+  const { petState } = useGameContext();
   const { food, happiness, energy } = petState;
   const { gameTick } = useGameLoop();
+  const { changeGame } = useGameHook();
   const { gameId } = useParams();
   const [updateGameData] = useMutation(UPDATE_GAMEDATA);
   const { loading, data } = useQuery(QUERY_USER, {
@@ -27,22 +29,7 @@ export default function GameDashboard() {
     // console.log(data.me.gameData);
     const gamesArray = data.me.gameData;
     const gameData = gamesArray.filter(({ _id }) => _id === gameId)[0];
-    // console.log(gameData);
-    setGameState((prev) => {
-      const newGame = { ...prev };
-      newGame.name = gameData.name;
-      newGame.gameId = gameId;
-      return newGame;
-    });
-    setPetState((prev) => {
-      // console.log('Pet', prev);
-      const newPet = { ...prev };
-      newPet.food = gameData.food;
-      newPet.energy = gameData.energy;
-      newPet.happiness = gameData.happiness;
-      // console.log('New Pet', newPet);
-      return newPet;
-    });
+    changeGame(gameData);
   }, [loading]);
 
   //Update data
@@ -63,6 +50,7 @@ export default function GameDashboard() {
 
   //Game Tick
   useEffect(() => {
+    if (food <= 0 || happiness <= 0 || energy <= 0) return;
     const interval = setInterval(() => {
       // console.log(petState);
