+++ client/src/lib/useGameHook.js
@@ -1,3 +1,4 @@
+import { useGameContext } from '../context/GameContext';
 import useGameLogic from './useGameLogic';
 import useGameLoop from './useGameLoop';
 
@@ -5,6 +6,7 @@ import useGameLoop from './useGameLoop';
 export default function useGameHook() {
   const { startGame } = useGameLoop();
   const { battle, executeAction } = useGameLogic();
+  const { setGameState, setPetState } = useGameContext();
 
   function play(happiness_array) {
     // warning(`${gameState.name} is playing!`);
@@ -29,5 +31,25 @@ export default function useGameHook() {
     startGame();
   }
 
-  return { play, feed, sleep, battleBeast, gameLoop };
+  function isAlive() {}
+
+  function changeGame(gameData) {
+    // console.log(gameData);
+    setGameState((prev) => {
+      const newGame = { ...prev };
+      newGame.name = gameData.name;
+      return newGame;
+    });
+    setPetState((prev) => {
+      // console.log('Pet', prev);
+      const newPet = { ...prev };
+      newPet.food = gameData.food;
+      newPet.energy = gameData.energy;
+      newPet.happiness = gameData.happiness;
+      // console.log('New Pet', newPet);
+      return newPet;
+    });
+  }
+
+  return { play, feed, sleep, battleBeast, gameLoop, changeGame };
 }
